requirements.txt
flet>=0.28.3
pyyaml
uiautomator2